api/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.10.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
bitarray==3.6.1
brotli==1.2.0
celery==5.5.3
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cssselect2==0.8.0
cytoolz==1.0.1
distro==1.9.0
Django==5.2.3
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-countries==7.6.1
django-filter==25.1
django-hosts==6.0
django-redis==6.0.0
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docstring_parser==0.17.0
drf-yasg==1.21.11
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
exceptiongroup==1.3.0
fonttools==4.61.1
gunicorn==23.0.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
kombu==5.5.4
mysql-connector-python==9.4.0
numpy==2.2.6
openai==1.99.9
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parsimonious==0.10.0
pillow==11.3.0
qrcode[pil]>=7.4
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.12.1
PyJWT==2.9.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==6.2.0
regex==2025.7.34
requests==2.32.4
rlp==4.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
toolz==1.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==67.0
webencodings==0.5.1
WooCommerce==3.0.0
zopfli==0.4.0
google-generativeai>=0.8.0
twilio>=9.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
